Migrate benchmarks, update workflows and docs
