+++ .github/workflows/lint-code.yml
@@ -12,5 +12,5 @@ jobs:
         uses: actions/setup-node@v2
         with:
           node-version: ${{ matrix.node-version }}
-      - run: make install
-      - run: make lint
+      - run: npm install
+      - run: npm run lint
